AP-11471: Assign own color to keras port objects
The color is the red from the Keras logo.
The general port object for Keras is the Keras..PortObjectBase. To
assign a color to this port object it needed to be registered via the
extension point. To register it via the extension point it needed to
have serializers. The serializers won't get called because the
serializer of the implementing classes will get called.

Also hidden is set to false for the specific port object
implementations. This will hide the port objects in metanode configs and
only show the general port object.